Moved utilities to plasma_core
