+++ manifest.in
@@ -1,0 +1,2 @@
+include *.md
+include LICENSE.txt
