+++ tests/root_chain/contracts/root_chain/test_root_chain.py
@@ -1,8 +1,8 @@
 import pytest
 import rlp
 from plasma.child_chain.transaction import Transaction, UnsignedTransaction
-from plasma.utils.merkle.fixed_merkle import FixedMerkle
-from plasma.utils.utils import confirm_tx, get_deposit_hash
+from plasma_core.utils.merkle.fixed_merkle import FixedMerkle
+from plasma_core.utils.utils import confirm_tx, get_deposit_hash
 
 
 @pytest.fixture
